ENG-487: fix inconsistencies
